Handle cases where multiple PRs refer to the same GitRef
If multiple PRs refer to the same GitRef, merging one with the
delete_branch flag automatically closes the others without merging them.
This is undesirable behavior, so check before we delete, and abort,
raising a exception if this is the case.

+++ github/PullRequest.py
@@ -55,7 +55,7 @@ import github.File
 import github.IssueComment
 import github.Commit
 import github.PullRequestReview
-
+import github.GithubException
 
 class PullRequest(github.GithubObject.CompletableGithubObject):
     """
@@ -775,7 +775,16 @@ class PullRequest(github.GithubObject.CompletableGithubObject):
             self.url + "/merge",
             input=post_parameters
         )
+
+
         if delete_branch == True:
+
+            # Check for other pull requests referencing the same ref.
+            # If found, abort.
+            remaining_pulls = self.head.repo.get_pulls(self.head.ref)
+            if remaining_pulls > 0:
+                raise GithubException.GitRefInUseException(data="PRs referencing this branch remain.  Not deleting the branch")
+
             self.head.repo.get_git_ref( "heads/%s"%(self.head.ref) ).delete()
